Added networking support (relatively major code change).
Now you can play over TCP/IP.  One player must be the server (listening
on a port) and the other the client (connecting on the same port to an
IP address).

It's a little awkward because the "network" player might be Player 2 on
one side of the connection and would necessarily be Player 1 on the
other side.

Things are a little clunky right now, and there are a number of bugs to
straighten out.

Oddly enough, it's possible for two computer players to play each other
over the network.