Changed tree growth rules so that:
1) Trees will grow counter-clockwise around cities
2) A tree that is blocked in all 8 directions by either cities or the
edge of the board will grow 4 new trees north, south, east, and west (or
as many as possible without growing past the edge of the board)

Trees no longer grow on grass or desert that is between two trees.

This growth rule puts a natural check on big blocks of cities.